Move operator source identity into context
Extend the harness source identity context to own operator and conversion-operator metadata names, bodies, and evidence.
